in the middle of refactoring. wip

+++ dlls/monsters.cpp
@@ -2216,7 +2216,7 @@ int CBaseMonster::IRelationship ( CBaseEntity *pTarget )
 	/*ABIOWEAPON*/	{ R_NO	,R_NO	,R_DL	,R_DL	,R_DL	,R_AL	,R_NO	,R_DL	,R_DL	,R_NO	,R_NO	,R_DL,	R_DL,	R_NO	}
 	};
 
-	return iEnemy[ Classify() ][ pTarget->Classify() ];
+	return iEnemy[ static_cast<int>(Classify()) ][ static_cast<int>(pTarget->Classify()) ];
 }
 
 //=========================================================
